requirements.txt
astroid==3.3.7
black==24.10.0
click==8.1.8
dill==0.3.9
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pylint==3.3.2
tomlkit==0.13.2
Flask==2.0.3
openai==0.27.0
requests==2.26.0
python-dotenv==0.21.0
moviepy==1.0.3
boto3==1.28.68

